fix: add retry logic for missing csrf token exceptions
