+++ apps/web/src/features/experience/api/create-experience.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { server } from '@/lib/server/client';
+import { server } from '@/lib/server';
 import { ExperienceType } from '@/types/experience';
 
 type CreateExperienceRequest = {
