+++ apps/web/src/features/message/api/get-conversations.ts
@@ -1,6 +1,6 @@
 import { UseQueryOptions, useQuery } from '@tanstack/react-query';
 
-import { server } from '@/lib/server/client';
+import { server } from '@/lib/server';
 import { Conversation } from '@/types/message';
 
 export type GetConversationsResponse = {
