+++ apps/web/src/features/message/api/get-messages.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { server } from '@/lib/server/client';
+import { server } from '@/lib/server';
 import { Message } from '@/types/message';
 import { url } from '@/util/server';
